app-container: try out "root-container" naming
Based on feeback from a number of folks the overloading of the
"container" has caused quite a bit of confusion. This PR tries
"root-container" on for size. Which if you think of cgroups as
heirarchy like the filesystem (which they are) makes a lot of s

This is an alternative to using a new term such as "pod", "grou
"cell". The idea is that we don't call `/` on the UNIX filesyst
anything special, we call it the "root-directory". So, lets see
can try this on for size here. There is a root-container and on
sub-containers. Apps execute as a direct sub-container of the
root-container.

https://github.com/coreos/rocket/issues/151