spec: default digest format is sha512
Currently the ACI spec leverages sha256 as the default digest format, which
is not as performant on 64-bit systems. A modern 64-bit processor can compute
a SHA-512 hash in about 50-75% of the time it takes to create a SHA-256 hash.
(This is because SHA-512 primarily uses 64-bit data types.) Considering that
container images can become rather large, this is substantial.

Fixes #8