schema: Fix optional app and required app.user and app.group.
From the SPEC app is optional. Make it a pointer to handle empty values and add
omitempty to App json tag to avoid creation of an "app":null json object.
Add a test to verify working empty app.
From the SPEC user and group are required. Fix and ddd tests.