*: split out VERSION and set during build time
This patch creates a single canonical file to represent the version of
the repo (spec + tooling) at a particular point in time, so we do not
have version information in multiple locations that needs to be kept in
sync. Build flags are then used (specifically, '-ldflags -X') to inject
this version into actool and the ACE validator. Finally, the `+git`
suffix is added to the current version to reflect the real state of
things and avoid confusion with the actual 0.1.1 release.