*: encode canonical spec version in VERSION
The versioning of the spec should be decoupled from that of the golang
schema and tooling (e.g. `actool` and the ACE validator). This patch
introduces a `VERSION` file in the root of the repo to encapsulate the
canonical spec version. For now, the golang version is tied to this
version (enforced through the `test` script), but in future they may
diverge (if for example the tooling moves to its own repo).

This change also introduces a `+git` suffix to the version to reflect
the real state of things and avoid confusion with the actual 0.1.1
release.

For more context, see the discussion in
https://github.com/appc/spec/pull/47