SPEC: have discovery walk up the tree
The basic idea is that people may not want to have a landing page for
every container name but will instead want a landing page for an overall
project. This would let people do this as the discovery algorithm will
walk up the tree from example.com/project/app1 to example.com/project to
example.com.