discovery: set default version label in doDiscover and not in NewAppFromString.
The app and labels serialization format parsed by NewAppFromString is not
strictly related to discovery/fetching but should also be used by other
functions (eg. asking an image to a local store without discovery and fetching).

Setting a default version label if it's missing shouldn't be done here but from
the functions that needs it (now in doDiscover).