schema: add "deviceFiles"
In order to add character device files or block device files in the ACI, one
could create them in the directory with mknod and let "actool build" pick them
up. But mknod requires root and it is desirable to be able to build the ACI
without being root.

This patch adds the specification for "deviceFiles":

{
    "acVersion": "0.1.1",
    "acKind": "ImageManifest",
    ...
    "deviceFiles": [
        "/dev/null",
        "/dev/zero",
        "/dev/net/tun"
    ]
}

Note that the device file type (character or block) and the major/minor are not
specified in the manifest: the device files will not be added in the ACI tar
archive by "actool build". Some device files use a dynamically allocated
major/minor so it might be different at build-time and at run-time.

stage1 must checks the device file type, major and minor of each file listed in
deviceFiles at run-time on the host and create them in the container root
filesystem.

appc/spec#99
appc/spec#127
coreos/rocket#389