SPEC: add mountPoints and remove fulfills
Add the concept of a mountPoint to the CRM and remove the idea that a volume
"fulfills" app mountPoints. The fulfills concept we troublesome because it was
putting the binding of a volume to a name at image/definition time instead of
runtime.

On the command line you could imagine it working something like this:

```
ace run
 example.com/webapp
 --volume ssl-share,type=empty
 example.com/ssl-key-downloader
 --mount ssl-share=example.com/ssl-key-downloader:ssl-out
 example.com/nginx-proxy
 --mount ssl-share=example.com/nginx-proxy:private-certs
```