spec: introduce pods
This is a first cut at pods, introducing a _Pod Manifest_ and tweaking
some of the other concepts slightly.

A Pod Manifest is simply a grouping of application images, and a
convenient deployable unit which can be provided as input to an executor.

However, the application image references in a pod manifest are not
necessarily deterministic: they may use the Dependency Matching
mechanism (as in the `dependencies` section of Image Manifest) to
resolve applications. For example, a pod manifest with an application
with a "version=latest" label might resolve to a different particular
application image at different points in time.

In contrast, the application references in a Container Runtime Manifest
MUST be deterministic (i.e. be image IDs).