spec: introduce Pods, remove CRMs
Following on from #150, this is the next iteration at an attempt to
introduce the concept of pods.

Following the most recent discussion on the topic, this patch attempts
to simplify things greatly by removing the ContainerRuntimeManifest
entirely and settling instead on a single PodManifest.

To achieve this, we give the PodManifest a dual purpose: it is both a
_deployable template_ (which can be used to stamp out instances of
pods) which before execution must be resolved into a final/reified
version. The allowed reified versions are a _subset_ of the template
versions; for example, app references in the reified version of a Pod
Manifest MUST be deterministic and contain image IDs (i.e. they cannot
rely on labels and dependency matching).

An important distinction from the previous CRM model is that a reified
PodManifest no longer _necessarily_ includes runtime information
specific to the executor (i.e., the UUID). It is hence possible that a
fully reified pod could be passed around as a deployable unit to
different executors.

This patch also goes to some length to eliminate the (frustratingly
ambiguous) word "container" from the spec, except for in qualified
contexts (e.g. "App Container Image").