aci: fix ManifestFromImage to handle ./manifest
The tar Header.Name (for example with gnu tar) can start with ./ (for example
./manifest). In this case ManifestFromImage will fail.

Plus, I noticed that the error was strange: unexpected error EOF. This is due
to the io.EOF error not being handled. So I rewrote the code to handle it (now
it's similar to github.com/coreos/rocket/pkg/tar.ExtractTar).