spec: enforce unique app names within a pod
Since 9aafd97 landed, app references within a pod gained a name field
that is distinct from the name of the underlying image. However, the
spec did not address whether this field needed to be unique, which made
it unclear as to how to facilitate e.g. multiple apps within a pod using
the same image.

This makes the name field a) required, and b) unique within a pod.

Accordingly, apps in a running pod can now be uniquely, globally
identified by the combination of Pod UUID + name.

Fixes #84