parse: use types.ACName instead of string for label's name.
This will avoid forced conversions from a map[string]string to
map[ACName]string.

Additionally it verifies that the provided name is a valid types.ACName

For rocket: when updating the vendored appc/spec to a version containing this
patch, rocket will require a fix to rkt/fetch_test.go.