schema: add armv7l/armv7b as valid archs for linux
After consulting with ARM engineers, it seems the most sensible approach
to architectures pre-aarch64 is to support the two endian variants of
ARMv7: `armv7l` for little-endian and `armv7b` for big-endian.

These also correspond to the output of `uname -m` on Linux systems.

This avoids the somewhat arbitrary armel/armhf/armv7 distinctions
(partly by not accounting for architectures older than ARMv7).