SPEC.md: proposal for numeric UID/GID in App user/group fields
This proposes a clarification of the handling of the "user" and "group" fields
and how the values should be handled within the execution environment.

Existing convention was that a user/group value of a numerical string would be
treated as the UID/GID. POSIX standard allows for a username to be all
numbers.

To account for this, the recommendation is that the user/group fields should
always be resolved with /etc/passwd or /etc/group first. If a valid entry is
not found, then it should treat the user/group string as the UID/GID if the
string is all numerical.

Fixes #231

References:

POSIX standard on usernames:
  http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_431
  http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_278