spec: fix gpg example and clarify ACI signing
As reported in #292, the example gpg command was truncating output by
re-using the input file. Instead, we write to a separate temporary
output file and then overwrite (mv) the original.

This also clarifies the ordering of operations for images: signing
should happen *after* (optional) encryption, not just compression. This
was previously discussed in #19, #156, and #204.