discovery: Update the package to expose the http.Client
This changes the HTTP handling within the discovery package to make its
http.Client an exported field. This has a couple of use cases where consumers
of the package may want to customize the handling, such as:

* Adjusting the dial handler with custom timeouts
* Altering the http.Transport on the client for testing outside of the
  discovery package.
* Altering the proxy used for discovery requests without setting altering the
  proxy settings for the whole process.

This also updates the internal test handling to use an internal interface that
is used to wrap http.Client.Get(). With this, the way the existing tests were
mocking out requests is preserved.