spec/schema: change annotations to ACIdentifier
https://github.com/appc/spec/commit/555e3e9b5e78781146eec8ff2cc30e5a27a0dafd
redefined ACName to a more restrictive syntax, and changed some fields
(like name, isolators, etc) to a new ACIdentifier type with lighter
restrictions similar to the previous ACName. Since annotations is a
field we expect to be namespaced it makes more sense for it to also use
the ACIdentifier type.