aci: make NewCompressedReader an io.ReadCloser
now NewCompressedReader returns an io.Reader but gzip.NewReader returns an
io.ReadCloser that was leaked.

This patch makes NewCompressedReader return an io.ReadCloser and wraps others
implementing only io.Reader making them an io.ReadCloser with a fake Close
method.

Consequently NewCompressedTarReader must implement a Close method needed to
close the underlying CompressedReader.

In the meantime duplicated code from actool/validate.go is removed in favor of
the one provided by the aci package.

In another patch also XzReader will be changed to become an io.ReadCloser.

As this is an api change, at least rkt and docker2aci needs to be patched when
updating vendored appc/spec in rkt.