aci: make XzReader implement io.ReadCloser
this is needed as the XzReader, now calling an external process, needs to be
closed or the underlying process will continue to read from the passed file
descriptor, creating a race with subsequent methods calling another XzReader.

The Close method closes the output pipe, kills the process and then waits for
it to exit.