spec: pod: mounts point to path instead of mountPoint
In order to support the use case of mounting at an arbitrary location
inside of the app rootfs we remove the "merge table" between app
mountPoint and volumes. Before, if someone wanted to simply mount over
say /var/log with a volume and that wasn't offered as a mountPoint by
the app the entire app section would need to be overridden.

This does not remove the mountPoint name or volume name so if a runtime
wishes to have this volume/mountPoint map it may still be implemented in
the runtime UX.

This concern was brought up by @thockin in #364.