schema: standardise on supplementaryGids
785b1ee5f39221f66884b416a9127e5fd0a7e31d introduced the ability for
users to set supplementary group IDs, but unfortunately they were
inconsistently referenced: in the spec itself as `supplementaryGids` but
in the schema code and (actool) as `supplementaryGroups`.

Since it is unambiguously (as a list of ints) a list of group IDs, this
changes the schema and actool to match the spec.