discovery: URL-encode label values
Appc label values can be arbitrary strings, so they might need to be
escaped to become a valid URL label.

The problem arose when the "version" label has a value like
"0.8.1+git734734", where the "+" is a reserved char as noted in RFC
3986, section 2.2.

The function still has some limitations wrt. label values, though.
Mention them in its documentation. I resisted the temptation of
recommending escaping the label values beforehand - if developer has
label values as separate strings then there is little point in using
NewAppFromString anyway. Also, to be frank, we are a bit smart about
equal signs (we split the part to a label and a value only by first
equal sign), but let's not advertise that - no need to bother
developer with unnecessary details.