spec: Define ACString type and use it for labels values
Having a label value being a freeform string is coumbersome to use and
query since it can contains non printable characters.
Also having it unlimited can bring to other problems.

This patch adds a ACString type that:

* is restricted to all unicode printable characters. Such characters
include letters, marks, numbers, punctuation, symbols, and the ASCII
space character, from unicode categories L, M, N, P, S and the ASCII
space character.
* it MUST be UTF-8 encoded
* it MUST be be 255 characters or less (not bytes).

This patch starts using this new type for labels values.

Some notes:

* There isn't a regexp to match the valid chars. In Go we can use
unicode.IsPrint or strconv.IsPrint

* This defines that the string MUST be UTF-8 (needed to correctly count
the number of characters) but this assumption is done everywhere (for
example json string marshall/unmarshall) when a string is involved. So
this should probably be defined at the top of the spec?