spec/ace: Specify Content-Type for metadata service.
Some of the metadata service endpoints explicitly specify a
Content-Type of text/plain; charset=us-ascii. For most other
endpoints, the ACE validator expects a Content-Type of
application/json.

The one exception is the HMAC verify endpoint which makes no mention
of its Content-Type nor even the contents of its response body. The
current implementation in `rkt` sets the Content-Type to
text/plain; charset=us-ascii which is now reflected in the spec.