spec: add os/linux/apparmor-profile isolator
AppArmor is a very widely used isolator in the GNU/Linux community,
providing an "easier to configure" alternative to SELinux. From the
perspective of AppC this is a fairly simple addition (profile data is
not included in the spec and must already be loaded into the kernel).

This feature is required for feature parity for ACI containers with OCI
ones (as AppArmor is also supported as a first-class citizen in OCI).

Signed-off-by: Aleksa Sarai <asarai@suse.de>